+++ Source/Krypton Components/Krypton.Toolkit/General/Definitions.cs
@@ -2298,46 +2298,6 @@ public enum ToolkitSupportType
 
 #endregion
 
-#region AboutBox Definitions
-
-#region Enum AboutToolkitPage
-
-internal enum AboutToolkitPage
-{
-    GeneralInformation = 0,
-    Discord = 1,
-    DeveloperInformation = 2,
-    Versions = 3
-}
-
-#endregion
-
-#region Enum AboutBoxFileInformationPage
-
-public enum AboutBoxFileInformationPage
-{
-    Application = 0,
-    Assemblies = 1,
-    AssemblyDetails = 2
-}
-
-#endregion
-
-#region Enum AboutBoxPage
-
-public enum AboutBoxPage
-{
-    GeneralInformation = 0,
-    Description = 1,
-    FileInformation = 2,
-    Theme = 3,
-    ToolkitInformation = 4
-}
-
-#endregion
-
-#endregion
-
 #region InformationBox Definitions
 
 #region Enum AutoCloseDefinedParameters